Update the projects files to the .NET Core format
As a first step to switch to .NET Core as the main runtime, convert the project files to NET Core format.

Note: many Analyzer warnings are currently disabled. The focus of the project is to get the existing extensions working again.

+++ VvvSampleTest/FolderCopyHookTest.cs
@@ -2,8 +2,8 @@
 //     Copyright (c) Victor Derks. See README.TXT for the details of the software licence.
 // </copyright>
 
-using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using MiniShellFramework.ComTypes;
 
 namespace VvvSampleTest
@@ -11,7 +11,7 @@ namespace VvvSampleTest
     /// <summary>
     /// This is a test class for FolderCopyHookTest and is intended
     /// to contain all FolderCopyHookTest Unit Tests
-    ///</summary>
+    /// </summary>
     [TestClass]
     public class FolderCopyHookTest
     {
@@ -28,7 +28,7 @@ namespace VvvSampleTest
         {
             var folderCopyHook = (ICopyHook)CreateFolderCopyHook();
             uint result = folderCopyHook.CopyCallback(new IntPtr(), FileOperation.Delete, 0, "bla", 0, "boe", 0);
-            Assert.AreEqual((uint)6, result);
+            Assert.AreEqual(6U, result);
         }
 
         [TestMethod]
